+++ client/capi/base/custom.hpp
@@ -3,7 +3,7 @@
 // Created Date: 03/11/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 21/11/2021
+// Last Modified: 22/11/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -32,7 +32,7 @@ class CustomGain final : public autd::core::Gain {
    */
   static autd::GainPtr create(const uint16_t* data, const size_t data_length) {
     const auto dev_num = (data_length + autd::NUM_TRANS_IN_UNIT - 1) / autd::NUM_TRANS_IN_UNIT;
-    std::vector<autd::DataArray> raw_data(dev_num);
+    std::vector<autd::core::GainData> raw_data(dev_num);
     for (size_t i = 0; i < dev_num; i++) {
       const auto rem = std::clamp(data_length - i * autd::NUM_TRANS_IN_UNIT, size_t{0}, autd::NUM_TRANS_IN_UNIT);
       std::memcpy(&raw_data[i][0], data + i * autd::NUM_TRANS_IN_UNIT, rem * sizeof(uint16_t));
@@ -40,7 +40,7 @@ class CustomGain final : public autd::core::Gain {
     return std::make_shared<CustomGain>(raw_data);
   }
   void calc(const autd::core::Geometry& geometry) override { this->_data = std::move(this->_raw_data); }
-  explicit CustomGain(std::vector<autd::DataArray> data) : Gain(), _raw_data(std::move(data)) {}
+  explicit CustomGain(std::vector<autd::core::GainData> data) : Gain(), _raw_data(std::move(data)) {}
   ~CustomGain() override = default;
   CustomGain(const CustomGain& v) noexcept = default;
   CustomGain& operator=(const CustomGain& obj) = default;
@@ -48,7 +48,7 @@ class CustomGain final : public autd::core::Gain {
   CustomGain& operator=(CustomGain&& obj) = default;
 
  private:
-  std::vector<autd::DataArray> _raw_data;
+  std::vector<autd::core::GainData> _raw_data;
 };
 
 /**