+++ client/include/autd3/core/gain.hpp
@@ -17,10 +17,21 @@
 #include "geometry.hpp"
 #include "hardware_defined.hpp"
 
-namespace autd::core {
+namespace autd {
+namespace core {
 
 class Gain;
 using GainPtr = std::shared_ptr<Gain>;
+
+struct Drive final {
+  Drive() : Drive(0x00, 0x00) {}
+  explicit Drive(const uint8_t duty, const uint8_t phase) : phase(phase), duty(duty) {}
+
+  uint8_t phase;
+  uint8_t duty;
+};
+
+using GainData = std::array<Drive, NUM_TRANS_IN_UNIT>;
 
 /**
  * @brief Gain controls the duty ratio and phase of each transducer in AUTD devices.
@@ -37,7 +48,7 @@ class Gain {
    * \param geometry Geometry
    */
   virtual void calc(const Geometry& geometry) {
-    for (size_t i = 0; i < geometry.num_devices(); i++) this->_data[i].fill(0x0000);
+    for (size_t i = 0; i < geometry.num_devices(); i++) this->_data[i].fill(Drive(0x00, 0x00));
   }
 
   /**
@@ -67,9 +78,8 @@ class Gain {
 
   /**
    * @brief Getter function for the data of duty ratio and phase of each transducers
-   * @details Each data is 16 bit unsigned integer, where high 8bits represents duty ratio and low 8bits represents phase
    */
-  [[nodiscard]] const std::vector<DataArray>& data() const { return _data; }
+  [[nodiscard]] const std::vector<GainData>& data() const { return _data; }
 
   Gain() noexcept : _built(false) {}
   virtual ~Gain() = default;
@@ -80,6 +90,7 @@ class Gain {
 
  protected:
   bool _built;
-  std::vector<DataArray> _data;
+  std::vector<GainData> _data;
 };
-}  // namespace autd::core
+}  // namespace core
+}  // namespace autd