+++ client/lib/gain/holo/array_fire/arrayfire_matrix.hpp
@@ -3,7 +3,7 @@
 // Created Date: 08/09/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 19/11/2021
+// Last Modified: 22/11/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -107,13 +107,13 @@ struct AFMatrix {
   }
   void solve(const std::shared_ptr<AFMatrix<T>>& b) { b->af_array = af::solve(af_array, b->af_array); }
   T dot(const std::shared_ptr<const AFMatrix<T>>& a) {
-    T v;
+    T v{};
     auto r = af::dot(af_array, a->af_array, AF_MAT_CONJ);
     r.host(&v);
     return v;
   }
   [[nodiscard]] double max_element() const {
-    T v;
+    T v{};
     (af::max)((af::max)(af_array)).host(&v);
     return std::abs(v);
   }
@@ -125,7 +125,7 @@ struct AFMatrix {
   }
 
   [[nodiscard]] T at(const size_t row, const size_t col) const {
-    T v;
+    T v{};
     af_array(static_cast<int>(row), static_cast<int>(col)).host(&v);
     return v;
   }
@@ -148,8 +148,8 @@ struct AFMatrix {
   void transfer_matrix(const double* foci, size_t foci_num, const std::vector<const double*>& positions, const std::vector<const double*>& directions,
                        double wavelength, double attenuation);
   void set_bcd_result(const std::shared_ptr<const AFMatrix<T>>& vec, size_t index);
-  void set_from_complex_drive(std::vector<core::DataArray>& dst, bool normalize, double max_coefficient);
-  void set_from_arg(std::vector<core::DataArray>& dst, size_t n);
+  void set_from_complex_drive(std::vector<core::GainData>& dst, bool normalize, double max_coefficient);
+  void set_from_arg(std::vector<core::GainData>& dst, size_t n);
   void back_prop(const std::shared_ptr<const AFMatrix<T>>& transfer, const std::shared_ptr<const AFMatrix<T>>& amps);
   void sigma_regularization(const std::shared_ptr<const AFMatrix<T>>& transfer, const std::shared_ptr<const AFMatrix<T>>& amps, double gamma);
   void col_sum_imag(const std::shared_ptr<AFMatrix<complex>>& src);
@@ -240,8 +240,8 @@ inline void AFMatrix<complex>::set_bcd_result(const std::shared_ptr<const AFMatr
   set(index, index, ii);
 }
 
-inline void AFMatrix<double>::set_from_complex_drive(std::vector<core::DataArray>&, const bool, const double) {}
-inline void AFMatrix<complex>::set_from_complex_drive(std::vector<core::DataArray>& dst, const bool normalize, const double max_coefficient) {
+inline void AFMatrix<double>::set_from_complex_drive(std::vector<core::GainData>&, const bool, const double) {}
+inline void AFMatrix<complex>::set_from_complex_drive(std::vector<core::GainData>& dst, const bool normalize, const double max_coefficient) {
   // FIXME: implement with ArrayFire
   const auto n = rows() * cols();
   const auto data = std::make_unique<complex[]>(n);
@@ -251,33 +251,31 @@ inline void AFMatrix<complex>::set_from_complex_drive(std::vector<core::DataArra
   size_t trans_idx = 0;
   for (size_t j = 0; j < n; j++) {
     const auto f_amp = normalize ? 1.0 : std::abs(data[j]) / max_coefficient;
-    const auto phase = core::utils::to_phase(std::arg(data[j]));
-    const auto duty = core::utils::to_duty(f_amp);
-    dst[dev_idx][trans_idx++] = core::utils::pack_to_u16(duty, phase);
-    if (trans_idx == core::NUM_TRANS_IN_UNIT) {
+    dst[dev_idx][trans_idx].duty = core::utils::to_duty(f_amp);
+    dst[dev_idx][trans_idx].phase = core::utils::to_phase(std::arg(data[j]));
+    if (++trans_idx == core::NUM_TRANS_IN_UNIT) {
       dev_idx++;
       trans_idx = 0;
     }
   }
 }
 
-inline void AFMatrix<double>::set_from_arg(std::vector<core::DataArray>& dst, const size_t n) {
+inline void AFMatrix<double>::set_from_arg(std::vector<core::GainData>& dst, const size_t n) {
   // FIXME: implement with ArrayFire
   size_t dev_idx = 0;
   size_t trans_idx = 0;
   const auto data = std::make_unique<double[]>(n);
   af_array(af::seq(static_cast<double>(n))).host(data.get());
   for (size_t j = 0; j < n; j++) {
-    constexpr uint8_t duty = 0xFF;
-    const auto phase = core::utils::to_phase(data[j]);
-    dst[dev_idx][trans_idx++] = core::utils::pack_to_u16(duty, phase);
-    if (trans_idx == core::NUM_TRANS_IN_UNIT) {
+    dst[dev_idx][trans_idx].duty = 0xFF;
+    dst[dev_idx][trans_idx].phase = core::utils::to_phase(data[j]);
+    if (++trans_idx == core::NUM_TRANS_IN_UNIT) {
       dev_idx++;
       trans_idx = 0;
     }
   }
 }
-inline void AFMatrix<complex>::set_from_arg(std::vector<core::DataArray>&, const size_t) {}
+inline void AFMatrix<complex>::set_from_arg(std::vector<core::GainData>&, const size_t) {}
 
 inline void AFMatrix<double>::back_prop(const std::shared_ptr<const AFMatrix<double>>&, const std::shared_ptr<const AFMatrix<double>>&) {}
 inline void AFMatrix<complex>::back_prop(const std::shared_ptr<const AFMatrix<complex>>& transfer,