+++ client/lib/gain/holo/cuda/cuda_matrix.hpp
@@ -3,7 +3,7 @@
 // Created Date: 10/09/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 11/09/2021
+// Last Modified: 22/11/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -24,7 +24,7 @@
 #include <memory>
 #include <vector>
 
-#include "autd3/core/hardware_defined.hpp"
+#include "autd3/core/gain.hpp"
 #include "autd3/gain/matrix.hpp"
 #include "matrix_pool.hpp"
 
@@ -90,8 +90,8 @@ struct CuMatrix {
   void transfer_matrix(const double* foci, size_t foci_num, const std::vector<const double*>& positions, const std::vector<const double*>& directions,
                        double wavelength, double attenuation);
   void set_bcd_result(const std::shared_ptr<const CuMatrix<T>>& vec, size_t index);
-  void set_from_complex_drive(std::vector<core::DataArray>& dst, bool normalize, double max_coefficient);
-  void set_from_arg(std::vector<core::DataArray>& dst, size_t n);
+  void set_from_complex_drive(std::vector<core::GainData>& dst, bool normalize, double max_coefficient);
+  void set_from_arg(std::vector<core::GainData>& dst, size_t n);
   void back_prop(const std::shared_ptr<const CuMatrix<T>>& transfer, const std::shared_ptr<const CuMatrix<T>>& amps);
   void sigma_regularization(const std::shared_ptr<const CuMatrix<T>>& transfer, const std::shared_ptr<const CuMatrix<T>>& amps, double gamma);
   void col_sum_imag(const std::shared_ptr<CuMatrix<complex>>& src);