+++ client/lib/gain/primitive.cpp
@@ -27,7 +27,8 @@ void Grouped::add(const size_t device_id, const GainPtr& gain) { this->_gain_map
 
 void Grouped::calc(const core::Geometry& geometry) {
   for (const auto& [_, g] : this->_gain_map) g->build(geometry);
-  for (size_t i = 0; i < geometry.num_devices(); i++) this->_data[i] = _gain_map.count(i) > 0 ? _gain_map[i]->data()[i] : DataArray{0x0000};
+  for (size_t i = 0; i < geometry.num_devices(); i++)
+    this->_data[i] = _gain_map.count(i) > 0 ? _gain_map[i]->data()[i] : core::GainData{core::Drive(0x00, 0x00)};
 }
 
 GainPtr PlaneWave::create(const Vector3& direction, const double amp) { return create(direction, core::utils::to_duty(amp)); }
@@ -42,8 +43,9 @@ void PlaneWave::calc(const core::Geometry& geometry) {
     size_t i = 0;
     for (const auto& trans_pos : dev) {
       const auto dist = trans_pos.dot(dir);
-      const auto phase = core::utils::to_phase(dist * wavenum);
-      this->_data[dev.id()][i++] = core::utils::pack_to_u16(this->_duty, phase);
+      this->_data[dev.id()][i].phase = core::utils::to_phase(dist * wavenum);
+      this->_data[dev.id()][i].duty = this->_duty;
+      i++;
     }
   }
 }
@@ -57,8 +59,9 @@ void FocalPoint::calc(const core::Geometry& geometry) {
     size_t i = 0;
     for (const auto& trans_pos : dev) {
       const auto dist = (trans_pos - this->_point).norm();
-      const auto phase = core::utils::to_phase(dist * wavenum);
-      this->_data[dev.id()][i++] = core::utils::pack_to_u16(this->_duty, phase);
+      this->_data[dev.id()][i].duty = this->_duty;
+      this->_data[dev.id()][i].phase = core::utils::to_phase(dist * wavenum);
+      i++;
     }
   }
 }
@@ -85,8 +88,9 @@ void BesselBeam::calc(const core::Geometry& geometry) {
       const auto r = trans_pos - this->_apex;
       const auto rr = rot * r;
       const auto d = std::sin(_theta_z) * std::sqrt(rr.x() * rr.x() + rr.y() * rr.y()) - std::cos(_theta_z) * rr.z();
-      const auto phase = core::utils::to_phase(d * wavenum);
-      this->_data[dev.id()][i++] = core::utils::pack_to_u16(this->_duty, phase);
+      this->_data[dev.id()][i].duty = this->_duty;
+      this->_data[dev.id()][i].phase = core::utils::to_phase(d * wavenum);
+      i++;
     }
   }
 }
@@ -98,6 +102,7 @@ GainPtr TransducerTest::create(const size_t transducer_index, const uint8_t duty
 void TransducerTest::calc(const core::Geometry& geometry) {
   size_t device_idx, local_trans_idx;
   core::Geometry::global_to_local_idx(_transducer_idx, &device_idx, &local_trans_idx);
-  this->_data[device_idx][local_trans_idx] = core::utils::pack_to_u16(this->_duty, this->_phase);
+  this->_data[device_idx][local_trans_idx].duty = this->_duty;
+  this->_data[device_idx][local_trans_idx].phase = this->_phase;
 }
 }  // namespace autd::gain